メモ関連のJSを切り出し
